- initialize RNG with seed - added time measurement

+++ src/main/scala/de/treufuss/filesync/Main.scala
@@ -1,14 +1,21 @@
 package de.treufuss.filesync
 
+import java.time.Duration
+
+import com.thedeanda.lorem.LoremIpsum
 import de.treufuss.filesync.filesystem.{FileSystem, FileSystemConf}
 import org.apache.logging.log4j.scala.Logging
-import com.thedeanda.lorem.LoremIpsum
 
 import scala.util.Random
 
 object Main extends App with Logging {
 
   logger.info("Program started")
+
+  //  val seed = Random.nextLong()
+  val seed = 0L
+
+  Random.setSeed(seed)
 
   val fsConf = FileSystemConf(
     pathSeparator = '/',
@@ -17,50 +24,59 @@ object Main extends App with Logging {
 
   val fs = new FileSystem[String](fsConf)
 
-  val lorem = LoremIpsum.getInstance
+  val lorem = new LoremIpsum(seed)
 
   def randomPathWithError = {
     val idSet = fs.idSet
     fs.pathOf(idSet.iterator.drop(Random.nextInt(idSet.size)).next).get + (if (Random.nextDouble() < 0.001) "a" else "")
   }
 
-  while (fs.size < 1000) {
+  def timed[R](block: => R): R = {
+    val t0 = System.nanoTime()
+    val result = block
+    val t1 = System.nanoTime()
+
+    val dur = Duration.ofNanos(t1 - t0)
+    println("Elapsed time: " + dur.toString.stripPrefix("PT"))
+    result
+  }
+
+  val sizeLimit = 1000
+  val operationLimit = 1000000
+
+  while (fs.size < sizeLimit) {
     val path = randomPathWithError
 
     fs.create(path, lorem.getWords(1))
   }
 
-  1 to 100 foreach { _ =>
+  timed {
+    1 to operationLimit foreach { iteration =>
 
-    Random.nextInt(4) match {
-      case 0 =>
-        while (fs.size < 1000) {
-          val path = randomPathWithError
+      if (iteration % (operationLimit / 10) == 0) println(iteration.toDouble / operationLimit * 100 + "%")
 
-          fs.create(path, lorem.getWords(1))
-        }
-      case 1 =>
-        1 to 5000 foreach { _ =>
-          val path = randomPathWithError
+      val path = randomPathWithError
+      val dest = randomPathWithError
 
+      Random.nextInt(4) match {
+        case 0 =>
+          if (fs.size < sizeLimit) {
+            fs.create(path, lorem.getWords(1))
+          }
+        case 1 =>
           fs.rename(path, lorem.getWords(1))
-        }
-      case 2 =>
-        1 to 5000 foreach { _ =>
-          val source = randomPathWithError
-          val dest = randomPathWithError
-
-          fs.move(source, dest)
-        }
-      case 3 =>
-        while (fs.size > 500) {
-          val path = randomPathWithError
-
-          fs.delete(path)
-        }
+        case 2 =>
+          if (!fs.move(path, dest)) fs.move(dest, path)
+        case 3 =>
+          if (fs.size > sizeLimit / 2) {
+            fs.delete(path)
+          }
+      }
     }
   }
 
-  println(fs)
-  println(fs.toJson)
+  println("seed: " + seed)
+
+  //  println(fs)
+  //  println(fs.toJson)
 }
